Fix stop-the-world race. (#108)
Previously the MMTk coordinator thread pretents to be the VM thread and
calls SafepointSynchronize::begin() directly.  This makes it race with
the real VM thread.

This commit introduces a dedicated "VM companion thread". The companion
therad requests the VM thread to execute a stop-the-world VM operation
(a VMOp_ThirdPartyHeapOperation), and the VM operation waits for the
stop-the-world GC to finish and lets the VM thread start the world
again.  This ensures the real VM thread is the only thread that calls
SafepointSynchronize::begin().  The MMTk binding initiates stack
scanning during stop-the-world.

Also made minor fixes to MMTKCollectorThread and MMTKContextThread

-   Removed is_VM_thread and is_GC_thread methods.
    -   It is not the VM thread.
    -   The `is_GC_thread` method is never used in the real VM thread,
        and is removed in the upstream in a later revision.
-   Initialize their native thread names on start.
    -   The native names are visible in GDB when debugging.
    -   Prepended  "MMTk" to their names to make them more obvious.