Fix resuming mutators.
There was bug that if a mutator calls `block_for_gc` after the GC
triggered stop-the-world, it will be block on MMTkHeap::_gc_lock before
reading mmtk_start_the_world_counter, and it will be waiting for a
non-existing GC in the future if the program ends soon.

We now read the mmtk_start_the_world_counter using atomic memory
operations before acquiring MMTkHeap::_gc_lock.  This avoids being
blocked by stop-the-world.