Check results for new CI, allow some benchmarks to fail (#211)
This PR adds a yaml file that records our expected results for DaCapo
tests. It also adds a script to check the results for each CI run. We
should generally expect a CI run to succeed (which means the results are
expected) once this PR is merged.

However, as we may have random failures that we haven't noticed, it is
possible a CI run still fails. In such a case, we should update the
`ci-expected-results.yml`, create an Github issue, and update
https://github.com/mmtk/mmtk-openjdk/issues/246.