Fix reference processing crash
Fixed a bug that when enqueuing references, it swapped the reference
pending list with the last node of the new linked list.  It should swap
with the *first* node instead.

Worked around a fact that the list given by the method
`ReferenceGlue::enqueue_references` may contain duplicated elements.
This is because the reference processor in mmtk-core was designed to
work with both of the two use patterns: (1) registering references when
created, and (2) discovering references while tracing.  Therefore, the
reference processor kept the list of Reference instances discovered in
the previous GC (in the from-space).  But as the OpenJDK binding scans
new Reference instances, it will add to-space addresses of the same
Reference instances to the reference processor.  After the references
are processed, both pointers will be normalized to the two-space copy,
but as two different entries.  We work around this problem in the
OpenJDK binding by simply skipping visited entries.