Acquire and release Heap_lock on VM_MMTkSTWOperation (#316)
This PR changes how we acquire `Heap_lock` for GC. With the changes, we
acquire the lock in `VM_MMTkSTWOperation::doit_prologue()`, and release
the lock in epilogue. We hold the lock during the entire STW. This is
the same behavior as OpenJDK's GCs.

We used to acquire the lock during reference enqueueing before
`swap_reference_pending_list()` (which requires the lock to be locked),
and acquire the lock before notifying the reference handler. This causes
deadlock as seen in #315 .
The `has_reference_pending_list` call used for an assertion in reference
enqueueing is removed, as it requires the current thread to hold the
lock which cannot be satisfied.

This PR should fix https://github.com/mmtk/mmtk-openjdk/issues/315.