Fix bugs related to `Reference.discovered` (#323)
Previously the object scanner does not scan this field when reference
processing is enabled, causing all nodes after the head node of any
linked lists of `Reference` instances linked by the `discovered` field
to be treat as garbage. We fix this by treating the `discovered` field
as a strong field.

Another bug is related to the deduplication procedure in
`mmtk_enqueue_references`. The code intended to handle adjacent
duplicated element, but was wrong. That bug was benign and the worst it
could do is making the last node temporarily point to itself before
re-assigned the proper pointer.