Warn instead of panic when `object_iterate` is called. (#325)
The` MMTkHeap` instance will receive calls to `safe_object_iterate` when
connected from VisualVM. If we use `guarantee(...)` to indicate the
function is not implemented, the process will abort. We now give a
warning instead. This does not properly implement the iteration, but
will allow the VM to be profiled with VisualVM.