Concurrent Immix (#311)
We added support for the new plan ConcurrentImmix added to the mmtk-core
in https://github.com/mmtk/mmtk-core/pull/1355

We implemented the SATB barrier fast paths in the OpenJDK binding, and
refactored the barriers to support both pre and post barriers, as well
as (weak) reference loading barrier. The OpenJDK binding is now aware of
concurrent marking, too.

---------

Co-authored-by: Yi Lin <qinsoon@gmail.com>
Co-authored-by: Kunshan Wang <wks1986@gmail.com>
Co-authored-by: mmtkgc-bot <mmtkgc.bot@gmail.com>