Demote re-registered nmethods (#336)
If an nmethod is in the mature roots and the nmethod is re-registered
due to code patching, the nmethod will appear in both the nursery roots
and the mature roots for one collection, causing some roots to be
scanned twice. This PR removes an nmethod from the mature roots, when
the nmethod is re-registered, so that the nmethod will only be present
in the nursery.